# api/swagger.yml
swagger: '2.0'
info:
  description: Diagnostics api
  version: 0.0.0
  title: HW Diag
basePath: /api
paths:
...
  /drives:
    get:
      tags:
        - drives
      summary: List all drives
      description: Returns all drives from `/dev/sd(.){1}` or `/dev/disks/by-path`
      operationId: getDrives
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              type: object
              properties:
                path:
                  type: string
                  example: "/dev/disks/by-path/xxxxxx"
                device:
                  type: string
                  example: "sda"
        '501':
          description: Error opening the directory
  /drives/fio:
    post:
      tags:
        - drives
      summary: Start fio run
      description: >-
        Run `fio` on the devices provided in the parameters. Creates a history
        item.
      operationId: runFioOnDrives
      parameters:
        - in: body
          name: body
          schema:
            type: object
            properties:
              devices:
                type: array
                items:
                  type: string
                description: 'List devices to test; `--filename` parameter of fio '
                example:
                  - /dev/sda
                  - /dev/sdb
              direct:
                type: integer
                description: '`--direct` parameter of fio'
                minimum: 0
                maximum: 1
                example: 0
              rw:
                type: string
                description: '`--rw` parameter of fio; default:''write'''
                example: write
              bs:
                type: string
                description: '`--bs` param of fio; blocksize'
                example: 4k
              runtime:
                type: integer
                description: Runtime in miliseconds `--runtime`
                example: 30
              numjobs:
                type: integer
                description: Number of processes to invoke
                example: 8
              name:
                type: string
                description: The name/title of the testrun
                example: etcher_test_2021_02_02_1011
              size:
                type: string
                description: Amount of data to use for the test
                example: 500m
              io_size:
                type: string
                description: '`--io_size parameter of fio`'
                example: 10g
              ioengine:
                type: string
                description: '`--ioengine parameter of fio`'
                example: libaio
              iodepth:
                type: number
                description: '`--iodepth parameter of fio`'
                example: 32
              fsync:
                type: number
                description: '`--fsync parameter of fio`'
                example: 10000
              invalidate:
                type: number
                description: '`--invalidate` parameter of fio'
                minimum: 0
                maximum: 1
                example: 0
              overwrite:
                type: number
                description: Overwrite anyting or not
                minimum: 0
                maximum: 1
                example: 0
              output_format:
                type: string
                example: json
      responses:
        '200':
          description: Success
        '501':
          description: Error
  /drives/sdk:
    post:
      tags:
        - drives
      summary: Start fio run
      description: >-
        Flashes random data to test the speed. Creates a history
        item.
      operationId: runSdkOnDrives
      parameters:
        - in: body
          name: body
          schema:
            type: object
            properties:
              devices:
                type: array
                items:
                  type: string
                description: 'List devices to test'
                example:
                  - /dev/sda
                  - /dev/sdb              
              size:
                type: number
                description: Amount of bytes to be written
                example: 500    
      responses:
        '200':
          description: Success
        '501':
          description: Error
  /drives/cancel:
    get:
      tags:
        - drives
      summary: Cancels all running tests
      description: Cancels all runnung tests
      operationId: cancelAllRun
      responses:
        '200':
          description: Success
        '501':
          description: Error, cannot kill process.
  /drives/fio/last:
    get:
      tags:
        - drives
      summary: Get last fio result
      description: Returns the result of the most recent `fio` run
      operationId: getLastFioResult
      responses:
        '200':
          description: Success
        '501':
          description: Error, no results yet.
  /drives/sdk/last:
    get:
      tags:
        - drives
      summary: Get last skd result
      description: Returns the result of the most recent run by etcher sdk
      operationId: getLastSdkResult
      responses:
        '200':
          description: Success
        '501':
          description: Error, no results yet.
...